chore: convert commons.color to ES Modules (#2152)
