+++ lib/commons/color/center-point-of-rect.js
@@ -1,14 +1,13 @@
-/* global color */
-
 /**
  * Get coordinates for an element's client rects or bounding client rect
  *
  * @method centerPointOfRect
  * @memberof axe.commons.color
  * @param {DOMRect} rect
+ * @deprecated
  * @returns {Object | undefined}
  */
-color.centerPointOfRect = function centerPointOfRect(rect) {
+function centerPointOfRect(rect) {
 	if (rect.left > window.innerWidth) {
 		return undefined;
 	}
@@ -27,4 +26,6 @@ color.centerPointOfRect = function centerPointOfRect(rect) {
 	);
 
 	return { x, y };
-};
+}
+
+export default centerPointOfRect;
