+++ lib/commons/color/color.js
@@ -1,5 +1,3 @@
-/* global color */
-
 /**
  * @class Color
  * @memberof axe.commons.color
@@ -8,7 +6,7 @@
  * @param {number} blue
  * @param {number} alpha
  */
-color.Color = function(red, green, blue, alpha) {
+function Color(red, green, blue, alpha) {
 	/** @type {number} */
 	this.red = red;
 
@@ -107,72 +105,6 @@ color.Color = function(red, green, blue, alpha) {
 
 		return 0.2126 * r + 0.7152 * g + 0.0722 * b;
 	};
-};
+}
 
-/**
- * Combine the two given color according to alpha blending.
- * @method flattenColors
- * @memberof axe.commons.color.Color
- * @instance
- * @param {Color} fgColor Foreground color
- * @param {Color} bgColor Background color
- * @return {Color} Blended color
- */
-color.flattenColors = function(fgColor, bgColor) {
-	var alpha = fgColor.alpha;
-	var r = (1 - alpha) * bgColor.red + alpha * fgColor.red;
-	var g = (1 - alpha) * bgColor.green + alpha * fgColor.green;
-	var b = (1 - alpha) * bgColor.blue + alpha * fgColor.blue;
-	var a = fgColor.alpha + bgColor.alpha * (1 - fgColor.alpha);
-
-	return new color.Color(r, g, b, a);
-};
-
-/**
- * Get the contrast of two colors
- * @method getContrast
- * @memberof axe.commons.color.Color
- * @instance
- * @param  {Color}  bgcolor  Background color
- * @param  {Color}  fgcolor  Foreground color
- * @return {number} The contrast ratio
- */
-color.getContrast = function(bgColor, fgColor) {
-	if (!fgColor || !bgColor) {
-		return null;
-	}
-
-	if (fgColor.alpha < 1) {
-		fgColor = color.flattenColors(fgColor, bgColor);
-	}
-
-	var bL = bgColor.getRelativeLuminance();
-	var fL = fgColor.getRelativeLuminance();
-
-	return (Math.max(fL, bL) + 0.05) / (Math.min(fL, bL) + 0.05);
-};
-
-/**
- * Check whether certain text properties meet WCAG contrast rules
- * @method hasValidContrastRatio
- * @memberof axe.commons.color.Color
- * @instance
- * @param  {Color}  bgcolor  Background color
- * @param  {Color}  fgcolor  Foreground color
- * @param  {number}  fontSize  Font size of text, in pixels
- * @param  {boolean}  isBold  Whether the text is bold
- * @return {{isValid: boolean, contrastRatio: number, expectedContrastRatio: number}}
- */
-color.hasValidContrastRatio = function(bg, fg, fontSize, isBold) {
-	var contrast = color.getContrast(bg, fg);
-	var isSmallFont =
-		(isBold && Math.ceil(fontSize * 72) / 96 < 14) ||
-		(!isBold && Math.ceil(fontSize * 72) / 96 < 18);
-	var expectedContrastRatio = isSmallFont ? 4.5 : 3;
-
-	return {
-		isValid: contrast > expectedContrastRatio,
-		contrastRatio: contrast,
-		expectedContrastRatio: expectedContrastRatio
-	};
-};
+export default Color;
