+++ lib/commons/color/element-has-image.js
@@ -1,4 +1,4 @@
-/* global color */
+import incompleteData from './incomplete-data';
 
 /**
  * Reports if an element has a background image or gradient
@@ -10,12 +10,12 @@
  * @param {Object|null} style
  * @return {Boolean}
  */
-color.elementHasImage = function elementHasImage(elm, style) {
+function elementHasImage(elm, style) {
 	const graphicNodes = ['IMG', 'CANVAS', 'OBJECT', 'IFRAME', 'VIDEO', 'SVG'];
 	const nodeName = elm.nodeName.toUpperCase();
 
 	if (graphicNodes.includes(nodeName)) {
-		axe.commons.color.incompleteData.set('bgColor', 'imgNode');
+		incompleteData.set('bgColor', 'imgNode');
 		return true;
 	}
 
@@ -26,11 +26,10 @@ color.elementHasImage = function elementHasImage(elm, style) {
 
 	if (hasBgImage) {
 		const hasGradient = /gradient/.test(bgImageStyle);
-		axe.commons.color.incompleteData.set(
-			'bgColor',
-			hasGradient ? 'bgGradient' : 'bgImage'
-		);
+		incompleteData.set('bgColor', hasGradient ? 'bgGradient' : 'bgImage');
 	}
 
 	return hasBgImage;
-};
+}
+
+export default elementHasImage;
