+++ lib/commons/color/element-is-distinct.js
@@ -1,4 +1,4 @@
-/* global color */
+import Color from './color';
 
 /**
  * Creates a string array of fonts for given CSSStyleDeclaration object
@@ -37,7 +37,7 @@ function elementIsDistinct(node, ancestorNode) {
 		result,
 		edge
 	) {
-		var borderClr = new color.Color();
+		var borderClr = new Color();
 		borderClr.parseRgbString(nodeStyle.getPropertyValue(edge + '-color'));
 
 		// Check if a border/outline was specified
@@ -85,4 +85,4 @@ function elementIsDistinct(node, ancestorNode) {
 	return hasStyle;
 }
 
-color.elementIsDistinct = elementIsDistinct;
+export default elementIsDistinct;
