+++ lib/commons/color/get-foreground-color.js
@@ -1,10 +1,14 @@
-/* global axe, color */
+import Color from './color';
+import getBackgroundColor from './get-background-color';
+import incompleteData from './incomplete-data';
+import flattenColors from './flatten-colors';
 
 function getOpacity(node) {
 	if (!node) {
 		return 1;
 	}
 
+	// TODO: es-module-utils.getNodeFromTree
 	const vNode = axe.utils.getNodeFromTree(node);
 
 	if (vNode && vNode._opacity !== undefined && vNode._opacity !== null) {
@@ -36,10 +40,10 @@ function getOpacity(node) {
  * @param {Color} bgColor
  * @return {Color|null}
  */
-color.getForegroundColor = function(node, noScroll, bgColor) {
+function getForegroundColor(node, noScroll, bgColor) {
 	const nodeStyle = window.getComputedStyle(node);
 
-	const fgColor = new color.Color();
+	const fgColor = new Color();
 	fgColor.parseRgbString(nodeStyle.getPropertyValue('color'));
 	const opacity = getOpacity(node);
 	fgColor.alpha = fgColor.alpha * opacity;
@@ -48,14 +52,16 @@ color.getForegroundColor = function(node, noScroll, bgColor) {
 	}
 
 	if (!bgColor) {
-		bgColor = color.getBackgroundColor(node, [], noScroll);
+		bgColor = getBackgroundColor(node, [], noScroll);
 	}
 
 	if (bgColor === null) {
-		const reason = axe.commons.color.incompleteData.get('bgColor');
-		axe.commons.color.incompleteData.set('fgColor', reason);
+		const reason = incompleteData.get('bgColor');
+		incompleteData.set('fgColor', reason);
 		return null;
 	}
 
-	return color.flattenColors(fgColor, bgColor);
-};
+	return flattenColors(fgColor, bgColor);
+}
+
+export default getForegroundColor;
