+++ lib/commons/color/get-own-background-color.js
@@ -1,4 +1,4 @@
-/* global color */
+import Color from './color';
 
 /**
  * Returns the non-alpha-blended background color of an element
@@ -9,8 +9,8 @@
  * @param {Object} elmStyle style of the element
  * @return {Color}
  */
-color.getOwnBackgroundColor = function getOwnBackgroundColor(elmStyle) {
-	const bgColor = new color.Color();
+function getOwnBackgroundColor(elmStyle) {
+	const bgColor = new Color();
 	bgColor.parseRgbString(elmStyle.getPropertyValue('background-color'));
 
 	if (bgColor.alpha !== 0) {
@@ -19,4 +19,6 @@ color.getOwnBackgroundColor = function getOwnBackgroundColor(elmStyle) {
 	}
 
 	return bgColor;
-};
+}
+
+export default getOwnBackgroundColor;
